Update Lambda Function code and Custom Resource.
Issue: A change in the parameter only updates the lambda function and never invokes the updated lambda function.
Fix: Update CFN resources. Pass the parameters to the Custom Resource and process them in the Lambda function as an event. This allows the change in the parameters to update the IAM policy.