Increase password length to match API tokens
When Jenkins is configure for another authentication provider (i.e. SAML), the existing admin login is invalidated. To work around this, an API token can be generated for the admin user. These tokens are 34 characters in length by default. It would be useful to be able to re-roll the stack to update the password in all of the scripts under this use-case.