Removing redundant packages.
The AMI used by the agent was customized to add things like Docker, git,
Java, etc. All references to installing these packages in the agent AMI
has been removed.

There are a few other things like ruby and a few gems that are also
installed. We should be adding this (along with the other packages
mentioned above) to our bakery (Packer/Ansible, etc) and rip out the
ruby and gems references in the CF template.